Load stashes before returning the context
