+++ src/modules/login/create.js
@@ -7,8 +7,9 @@ import { base64 } from '../../util/encoding.js'
 import type { ApiInput } from '../root.js'
 import { authRequest } from './authServer.js'
 import { makeKeysKit } from './keys.js'
+import { fixUsername, hashUsername } from './login-selectors.js'
 import type { LoginKit, LoginTree } from './login-types.js'
-import { fixUsername, hashUsername, saveStash } from './loginStore.js'
+import { saveStash } from './loginStore.js'
 import { makePasswordKit } from './password.js'
 import { makeChangePin2Kit } from './pin2.js'
