+++ src/modules/root-pixie.js
@@ -10,8 +10,6 @@ import type { CurrencyOutput } from './currency/currency-pixie.js'
 import currency from './currency/currency-pixie.js'
 import type { ExchangeOutput } from './exchange/exchange-pixie.js'
 import exchange from './exchange/exchange-pixie.js'
-import type { LoginOutput } from './login/login-pixie.js'
-import login from './login/login-pixie.js'
 import type { ScryptOutput } from './scrypt/scrypt-pixie.js'
 import scrypt from './scrypt/scrypt-pixie.js'
 
@@ -21,7 +19,6 @@ export type RootOutput = {
   +context: ContextOutput,
   +currency: CurrencyOutput,
   +exchange: ExchangeOutput,
-  +login: LoginOutput,
   +scrypt: ScryptOutput
 }
 
@@ -30,6 +27,5 @@ export const rootPixie = combinePixies({
   context,
   currency,
   exchange,
-  login,
   scrypt
 })
